refactor(#133): useTagInput 훅으로 분리하기
