+++ apps/frontend/src/shared/common/component/index.ts
@@ -1,3 +1,2 @@
 export * from './ModalBox';
-export * from './TagInput';
 export * from './Time';
