+++ apps/frontend/src/shared/tagInput/TagInput.tsx
@@ -1,6 +1,6 @@
-import { useRef, useState } from 'react';
 import { Badge, Button, Input, Text } from '@froxy/design/components';
 import { TiDelete } from 'react-icons/ti';
+import { useTagInput } from './useTagInput';
 
 interface TagInputProp {
   value: string[];
@@ -8,33 +8,7 @@ interface TagInputProp {
 }
 
 export function TagInput({ value, onChange }: TagInputProp) {
-  const inputRef = useRef<HTMLInputElement>(null);
-  const [isComposing, setIsComposing] = useState(false);
-
-  const addTag = (tagValue: string) => {
-    if (tagValue.trim()) {
-      onChange([...value, tagValue.trim()]);
-
-      if (inputRef.current?.value) inputRef.current.value = '';
-    }
-  };
-
-  const onInputKeyDown = (event: React.KeyboardEvent<HTMLInputElement>) => {
-    // Enter 키 입력 시
-    if (event.key === 'Enter') {
-      event.preventDefault();
-
-      // 입력 중이 아닐 때만 태그 추가
-      if (!isComposing && inputRef.current) {
-        addTag(inputRef.current.value);
-      }
-    }
-  };
-
-  const removeTag = (index: number) => {
-    const updated = value.filter((_, i) => index !== i);
-    onChange(updated);
-  };
+  const { inputRef, setIsComposing, onInputKeyDown, removeTag } = useTagInput(value, onChange);
 
   return (
     <div>
