+++ apps/frontend/src/shared/tagInput/index.ts
@@ -1,0 +1,1 @@
+export * from './TagInput';
