+++ apps/frontend/src/widget/lotusUpdate/LotusUpdateForm.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { Button, Heading, Input, Text } from '@froxy/design/components';
 import { useLotusSuspenseQuery } from '@/feature/lotus';
-import { TagInput } from '@/shared';
+import { TagInput } from '@/shared/tagInput';
 
 interface LotusUpdateFormProps {
   lotusId: string;
